pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask==2.3.3",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "python-telegram-bot==20.7",
    "requests>=2.32.3",
    "telegram>=0.0.1",
    "tlgbotfwk>=0.4.61",
    "vk-api>=11.9.9",
]
